+++ app/scripts/lizard-nxt.js
@@ -242,39 +242,28 @@ angular.module('lizard-nxt')
     }
   });
 
+  // TIME MODEL
   var now = Date.now(),
       hour = 60 * 60 * 1000;
 
-  // TIME MODEL
   $scope.timeState = {
     start: now - 4 * hour,
     end: now + hour,
+    at: now - 2 * hour,
     changedZoom: Date.now(),
     zoomEnded: null,
+    aggWindow: 1000 * 60 * 5,
     animation: {
-      start: undefined,
       playing: false,
       enabled: false,
-      currentFrame: 0,
-      lenght: 0,
-      minLag: 50, // Time in ms between frames
-      stepSize: 1000
     }
   };
-  // initialise 'now'
-  $scope.timeState.at = $scope.timeState.start;
-  $scope.timeState.animation.start = $scope.timeState.start;
-  // get time resolution in ms per pixel
-  $scope.timeState.resolution =
-    ($scope.timeState.end - $scope.timeState.start) / window.innerWidth;
 
-  $scope.timeState.aggWindow = UtilService.getAggWindow($scope.timeState.start,
-                                                        $scope.timeState.end,
-                                                        window.innerWidth);
-
-  $scope.timeState.setTimeStateBuffering = function (buffering) {
-    $scope.timeState.buffering = buffering;
-   };
+  $scope.timeState.aggWindow = UtilService.getAggWindow(
+    $scope.timeState.start,
+    $scope.timeState.end,
+    window.innerWidth
+  );
   // END TIME MODEL
 
   /**